try JanK suggestion and use LoadLibrary instead of Assembly.LoadFile
