+++ src/libraries/System.Diagnostics.Process/tests/ProcessModuleTests.cs
@@ -1,16 +1,13 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices;
 using Microsoft.DotNet.RemoteExecutor;
 using Xunit;
 
 namespace System.Diagnostics.Tests
 {
-    public class ProcessModuleTests : ProcessTestBase
+    public partial class ProcessModuleTests : ProcessTestBase
     {
         [Fact]
         public void TestModuleProperties()
@@ -91,42 +88,6 @@ namespace System.Diagnostics.Tests
 
             process.Dispose();
             Assert.Equal(expectedCount, disposedCount);
-        }
-
-        public static bool Is_LongModuleFileNamesAreSupported_TestEnabled
-            => OperatingSystem.IsWindows() // it's specific to Windows
-            && PathFeatures.AreAllLongPathsAvailable() // we want to test long paths
-            && !PlatformDetection.IsMonoRuntime // Assembly.LoadFile used the way this test is implemented fails on Mono
-            && RuntimeInformation.ProcessArchitecture != Architecture.X86; // for some reason it's flaky on x86
-
-        [ConditionalFact(typeof(ProcessModuleTests), nameof(Is_LongModuleFileNamesAreSupported_TestEnabled))]
-        public void LongModuleFileNamesAreSupported()
-        {
-            // To be able to test Long Path support for ProcessModule.FileName we need a .dll that has a path > 260 chars.
-            // Since Long Paths support can be disabled (see the ConditionalFact attribute usage above),
-            // we just copy "LongName.dll" from bin to a temp directory with a long name and load it from there.
-            // Loading from new path is possible because the type exposed by the assembly is not referenced in any explicit way.
-            const string libraryName = "LongPath.dll";
-            const int minPathLength = 261;
-
-            string testBinPath = Path.GetDirectoryName(typeof(ProcessModuleTests).Assembly.Location);
-            string libraryToCopy = Path.Combine(testBinPath, libraryName);
-            Assert.True(File.Exists(libraryToCopy), $"{libraryName} was not present in bin folder '{testBinPath}'");
-
-            string directoryWithLongName = Path.Combine(TestDirectory, new string('a', Math.Max(1, minPathLength - TestDirectory.Length)));
-            Directory.CreateDirectory(directoryWithLongName);
-
-            string longNamePath = Path.Combine(directoryWithLongName, libraryName);
-            Assert.True(longNamePath.Length > minPathLength);
-
-            File.Copy(libraryToCopy, longNamePath);
-            Assert.True(File.Exists(longNamePath));
-
-            Assembly loaded = Assembly.LoadFile(longNamePath);
-            Assert.Equal(longNamePath, loaded.Location);
-
-            string[] modulePaths = Process.GetCurrentProcess().Modules.Cast<ProcessModule>().Select(module => module.FileName).ToArray();
-            Assert.Contains(longNamePath, modulePaths);
         }
     }
 }
